perl.h: Use auxiliary #define for simplicity
Prior to this commit we avoided #defining a symbol until later, so that
that fact could be used for further decisions.  But this is a bit
awkward, and so this commit introduces an additional #define that can be
used for the decisions, while defining the symbol at the place where it
makes the most sense to.