regen/reentr.pl: Redefine lock if redefine function
The generated header may define a macro that overrides a libc function
call with an equivalent reentrant one.  That means that the mutex lock
macros for that function are no longer correct; they are for the
original.  This commit causes them to be changed to the revised one.
Typically the result will be less strict locking.