Avoid deadlock if request a read lock while holding a write one
iIn a reentrant mutex, the previous commit added the ability to have
multiple readers.  But if a thread owns a mutex, and requests a read
lock as well, the thread will be deadlocked by itself.

Avoid this by testing for the condition and handling separately.

All reentrant mutexes currently have no read lock attempts.  So this
case is entirely theoretical for the moment.